// src/@types/models/create-brand-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1.0.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 *
 *
 * @export
 * @interface CreateBrandDto
 */
export interface CreateBrandDto {
  /**
   * @type {Blob}
   * @memberof CreateBrandDto
   */
  icon: any;

  /**
   * @type {string}
   * @memberof CreateBrandDto
   */
  title: string;
}